+++ lib/parse-args.js
@@ -1,11 +1,12 @@
 const defaultExclude = require('@istanbuljs/schema/default-exclude')
 const defaultExtension = require('@istanbuljs/schema/default-extension')
 const findUp = require('find-up')
-const { readFileSync } = require('fs')
 const Yargs = require('yargs/yargs')
 const { applyExtends } = require('yargs/helpers')
 const parser = require('yargs-parser')
-const { resolve, extname, basename } = require('path')
+const { resolve } = require('path')
+
+const { CONFIG_FILE_NAMES, loadConfigFile } = require('./load-config')
 
 function buildYargs (withCommands = false) {
   const yargs = Yargs([])
@@ -18,7 +19,7 @@ function buildYargs (withCommands = false) {
         const config = loadConfigFile(path)
         return applyExtends(config, process.cwd(), true)
       },
-      default: () => findUp.sync(getConfigFileNames())
+      default: () => findUp.sync(CONFIG_FILE_NAMES)
     })
     .option('reporter', {
       alias: 'r',
@@ -194,52 +195,6 @@ function buildYargs (withCommands = false) {
   return yargs
 }
 
-function loadConfigFile (path) {
-  let config = {}
-  const jsExts = ['.js', '.cjs']
-  const ymlExts = ['.yml', '.yaml']
-  const fileName = basename(path)
-
-  const ext = extname(path).toLowerCase()
-  if (jsExts.includes(ext)) {
-    config = require(path)
-  } else if (ymlExts.includes(ext)) {
-    config = require('js-yaml').load(readFileSync(path, 'utf8'))
-  } else if (ext === '.json' || fileName.slice(-2) === 'rc') {
-    config = JSON.parse(readFileSync(path, 'utf8'))
-  }
-
-  // Should the process die if none of these filename extensions are found?
-  if (Object.keys(config).length === 0) {
-    throw new Error(`Unsupported file type ${ext} while reading file ${path}`)
-  }
-
-  return config
-}
-
-function getConfigFileNames () {
-  return [
-    '.c8rc',
-    '.c8rc.json',
-    '.c8rc.yml',
-    '.c8rc.yaml',
-    '.c8rc.js',
-    '.c8rc.cjs',
-    '.c8.config.js',
-    '.c8.config.cjs',
-    'c8.config.js',
-    'c8.config.cjs',
-    '.nycrc',
-    '.nycrc.json',
-    '.nycrc.yml',
-    '.nycrc.yaml',
-    '.nyc.config.js',
-    '.nyc.config.cjs',
-    'nyc.config.js',
-    'nyc.config.cjs'
-  ]
-}
-
 function hideInstrumenterArgs (yargv) {
   let argv = process.argv.slice(1)
   argv = argv.slice(argv.indexOf(yargv._[0]))
@@ -266,7 +221,5 @@ function hideInstrumenteeArgs () {
 module.exports = {
   buildYargs,
   hideInstrumenterArgs,
-  hideInstrumenteeArgs,
-  getConfigFileNames,
-  loadConfigFile
+  hideInstrumenteeArgs
 }
