+++ test/load-config-helper.js
